processArchive: use scrutinizer and add initial scan
Instead of subscribe directly to the vochain events, subscribe to the
scrutinizer, this way we have access to a more standarized Process
details and the full results (including weight).

In addition perform an initial scan over all existing processes to
ensure all of them are saved into the process archive.

Signed-off-by: p4u <pau@dabax.net>